Moved some things to the shared impl
