+++ queue/src/fiskaltrust.Middleware.Localization.QueueGR/GRSSCD/AADE/AADEFactory.cs
@@ -5,7 +5,6 @@ using System.Web;
 using System.Xml.Serialization;
 using fiskaltrust.Api.POS.Models.ifPOS.v2;
 using fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.AADE.Models;
-using fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.myDataSCU;
 using fiskaltrust.Middleware.Localization.QueueGR.Models.Cases;
 using fiskaltrust.Middleware.Localization.v2.Helpers;
 using fiskaltrust.Middleware.Localization.v2.Models.ifPOS.v2.Cases;
@@ -23,33 +22,6 @@ public class AADEFactory
         _masterDataConfiguration = masterDataConfiguration;
     }
 
-    public void ValidateReceiptRequest(ReceiptRequest receiptRequest)
-    {
-        if (receiptRequest.cbChargeItems.Any(x => x.ftChargeItemCase.IsTypeOfService(ChargeItemCaseTypeOfService.NotOwnSales)) && !receiptRequest.cbChargeItems.All(x => x.ftChargeItemCase.IsTypeOfService(ChargeItemCaseTypeOfService.NotOwnSales)))
-        {
-            throw new Exception("It is not allowed to mix agency and non agency receipts.");
-        }
-
-        if (receiptRequest.cbChargeItems.Sum(x => x.Amount) != receiptRequest.cbPayItems.Sum(x => x.Amount))
-        {
-            throw new Exception("The sum of the charge items must be equal to the sum of the pay items.");
-        }
-
-        if(receiptRequest.ftReceiptCase.IsFlag(ReceiptCaseFlags.Void))
-        {
-            throw new Exception("The Voiding of documents is not supported. Please use refund.");
-        }
-
-        if (receiptRequest.ftReceiptCase.IsFlag(ReceiptCaseFlagsGR.IsSelfPricingOperation))
-        {
-            throw new Exception("SelfPricing is not supported.");
-        }
-
-        if (AADEMappings.RequiresCustomerInfo(AADEMappings.GetInvoiceType(receiptRequest)) && !receiptRequest.ContainsCustomerInfo())
-        {
-            throw new Exception("Customer info is required for this invoice type");
-        }
-    }
 
     public InvoicesDoc MapToInvoicesDoc(List<ftQueueItem> queueItems)
     {
@@ -82,7 +54,7 @@ public class AADEFactory
 
     public InvoicesDoc MapToInvoicesDoc(ReceiptRequest receiptRequest, ReceiptResponse receiptResponse)
     {
-        ValidateReceiptRequest(receiptRequest);
+        MyDataAADEValidation.ValidateReceiptRequest(receiptRequest);
         var inv = CreateInvoiceDocType(receiptRequest, receiptResponse);
         var doc = new InvoicesDoc
         {
