--- queue/src/fiskaltrust.Middleware.Localization.QueueGR/GRSSCD/myDataSCU/MyDataPaymentMethods.cs
@@ -1,14 +1,0 @@
-﻿#pragma warning disable
-namespace fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.myDataSCU;
-
-public class MyDataPaymentMethods
-{
-    public const int DomesticPaymentAccount = 1;
-    public const int ForeignPaymentsSpecialAccount = 2;
-    public const int Cash = 3;
-    public const int Cheque = 4;
-    public const int OnCredit = 5;
-    public const int WebBanking = 6;
-    public const int PosEPos = 7;
-    public const int IrisDirectPayments = 8;
-}
