+++ queue/src/fiskaltrust.Middleware.Localization.v2/Helpers/ReceiptRequestExtensions.cs
@@ -1,9 +1,6 @@
 ﻿using fiskaltrust.Api.POS.Models.ifPOS.v2;
-using fiskaltrust.Middleware.Localization.v2.Models.ifPOS.v2.Cases;
-using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
+using fiskaltrust.Middleware.Localization.v2.Models;
+using System.Text.Json;
 
 namespace fiskaltrust.Middleware.Localization.v2.Helpers;
 
@@ -32,5 +29,39 @@ public static class ReceiptRequestExtensions
             }
         }
         return data;
+    }
+
+    public static bool ContainsCustomerInfo(this ReceiptRequest receiptRequest)
+    {
+        if (receiptRequest.cbCustomer != null)
+        {
+            return JsonSerializer.Deserialize<MiddlewareCustomer>(JsonSerializer.Serialize(receiptRequest.cbCustomer), new JsonSerializerOptions
+            {
+                PropertyNameCaseInsensitive = true
+            }) != null;
+        }
+        return false;
+    }
+
+    public static MiddlewareCustomer? GetCustomerOrNull(this ReceiptRequest receiptRequest)
+    {
+        if (receiptRequest.cbCustomer != null)
+        {
+            return JsonSerializer.Deserialize<MiddlewareCustomer>(JsonSerializer.Serialize(receiptRequest.cbCustomer), new JsonSerializerOptions
+            {
+                PropertyNameCaseInsensitive = true
+            });
+        }
+        return null;
+    }
+
+    public static bool HasGreeceCountryCode(this ReceiptRequest receiptRequest)
+    {
+        return ((ulong) receiptRequest.ftReceiptCase & 0xFFFF_0000_0000_0000) == 0x4752_0000_0000_0000;
+    }
+
+    public static bool HasNonEUCountryCode(this ReceiptRequest receiptRequest)
+    {
+        return ((ulong) receiptRequest.ftReceiptCase & 0xFFFF_0000_0000_0000) == 0x0000_0000_0000_0000;
     }
 }
